Extracted search query validation from domain
