+++ features/search/build.gradle.kts
@@ -26,4 +26,5 @@ dependencies {
     testImplementation("junit:junit:4.13.2")
     testImplementation("com.google.truth:truth:1.1.3")
     testImplementation("app.cash.turbine:turbine:0.5.2")
+    testImplementation("com.github.ubiratansoares:burster:0.1.1")
 }
